requirements.txt
pandas>=1.3.5
numpy>=1.20.0
requests>=2.27.1
beautifulsoup4>=4.10.0
scikit-learn>=1.0.2
matplotlib>=3.5.1
seaborn>=0.11.2
pybaseball>=2.2.5
python-dotenv>=0.20.0